+++ tests/logging_test.py
@@ -14,7 +14,8 @@ except ImportError:
     HAS_GEVENT = False
 
 from tests import TestLogHandler
-from tests.actor_test import EarlyFailingActor, LateFailingActor
+from tests.actor_test import (
+    EarlyFailingActor, LateFailingActor, FailingOnFailureActor)
 
 
 class LoggingNullHandlerTest(unittest.TestCase):
@@ -114,6 +115,19 @@ class ActorLoggingTest(object):
             'Unhandled exception in %s:' % actor_ref,
             log_record.getMessage())
 
+    def test_exception_in_on_failure_is_logged(self):
+        self.log_handler.reset()
+        failure_event = self.event_class()
+        actor_ref = self.FailingOnFailureActor.start(failure_event)
+        actor_ref.tell({'command': 'raise exception'})
+        failure_event.wait(5)
+        time.sleep(0.01)  # Too ensure that the log handler is updated
+        self.assertEqual(2, len(self.log_handler.messages['error']))
+        log_record = self.log_handler.messages['error'][0]
+        self.assertEqual(
+            'Unhandled exception in %s:' % actor_ref,
+            log_record.getMessage())
+
 
 class AnActor(object):
     def __init__(self, on_stop_was_called, on_failure_was_called):
@@ -151,6 +165,9 @@ class ThreadingActorLoggingTest(ActorLoggingTest, unittest.TestCase):
     class LateFailingActor(LateFailingActor, ThreadingActor):
         pass
 
+    class FailingOnFailureActor(FailingOnFailureActor, ThreadingActor):
+        pass
+
 
 if HAS_GEVENT:
     class GeventActorLoggingTest(ActorLoggingTest, unittest.TestCase):
@@ -163,4 +180,7 @@ if HAS_GEVENT:
             pass
 
         class LateFailingActor(LateFailingActor, GeventActor):
+            pass
+
+        class FailingOnFailureActor(FailingOnFailureActor, GeventActor):
             pass